test(host): split Beads context path coverage
